requirements.txt
Django==5.2
djangorestframework==3.15.2
drf-yasg==1.21.15
django-cors-headers==4.9.0
mysqlclient==2.2.8
python-decouple==3.8
paho-mqtt==1.6.1
